nyt_etl/requirements.txt
astroid==2.14.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
dnspython==2.3.0
feedparser==6.0.10
idna==3.4
isort==5.12.0
joblib==1.2.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
nltk==3.7
numpy==1.24.2
pandas==1.5.3
platformdirs==3.0.0
pylint==2.16.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
schedule==1.1.0
sgmllib3k==1.0.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing-extensions==4.5.0
urllib3==1.26.14
wrapt==1.14.1
